Xoá luôn tham chiếu khu vực ở điều kiện khi thực hiện xoá khu vực

+++ web/src/view/checkins/components/area/index.vue
@@ -1,55 +1,69 @@
 <template>
-    <div>
-        <div class="p-1 my-1">
-            <el-button type="primary" icon="plus" @click="addNewArea()">Thêm khu vực</el-button>
-        </div>
-
-        <el-table :data="tableData" style="width: 100%">
-            <el-table-column prop="Area.name" label="Nhóm"></el-table-column>
-            <el-table-column prop="Area.latitude" label="Kinh độ"></el-table-column>
-            <el-table-column prop="Area.longitude" label="Vĩ độ"></el-table-column>
-            <el-table-column prop="" label="Bán kính">
-                <template #default="scope">
-                    {{scope.row.radius ?? scope.row.Area.radius}}
-                </template>
-            </el-table-column>
-            <el-table-column label="Hành động">
-                <template #default="scope">
-                    <el-button type="danger" plain round @click="onDeleteArea(scope.row.ID)">Xoá</el-button>
-                </template>
-            </el-table-column>
-        </el-table>
-        <div class="flex justify-end">
-            <el-pagination v-model:current-page="page" v-model:page-size="pageSize" :page-sizes="[20, 50, 100, 500]"
-                :size="size" :background="true" layout="total, sizes, prev, pager, next, jumper" :total="total"
-                @size-change="handleSizeChange" @current-change="handleCurrentChange" />
-        </div>
-        <el-drawer destroy-on-close size="800" v-model="dialogFormVisible" :show-close="false"
-            :before-close="closeDialog">
-            <template #header>
-                <div class="flex justify-between items-center">
-                    <span class="text-lg">{{ type === 'create' ? 'Thêm mới' : 'Chỉnh sửa' }}</span>
-                    <div>
-                        <el-button type="primary" @click="enterDialog">Đồng ý</el-button>
-                        <el-button @click="dialogFormVisible = false">Hủy</el-button>
-                    </div>
-                </div>
-            </template>
-
-            <el-form :model="formData" label-position="top" :rules="rules" ref="elFormRef" label-width="80px">
-                <el-form-item label="Tên khu vực:" prop="areaId">
-                    <el-select v-model="formData.areaId" placeholder="Chọn khu vực" clearable filterable
-                        @change="onSelectChange">
-                        <el-option v-for="item in areaOptions" :key="item.ID" :label="item.name" :value="item.ID">
-                        </el-option>
-                    </el-select>
-                </el-form-item>
-                <el-form-item label="Bán kính:" prop="radius">
-                    <el-input-number v-model="formData.radius" style="width:100%" :precision="2" :clearable="true" />
-                </el-form-item>
-            </el-form>
-        </el-drawer>
+  <div>
+    <div class="p-1 my-1">
+      <el-button type="primary" icon="plus" @click="addNewArea()">
+        Thêm khu vực
+      </el-button>
     </div>
+
+    <el-table :data="tableData" style="width: 100%" border>
+      <el-table-column type="index" label="STT" width="60" />
+      <el-table-column prop="Area.name" label="Nhóm" />
+      <el-table-column prop="Area.latitude" label="Kinh độ" />
+      <el-table-column prop="Area.longitude" label="Vĩ độ" />
+      <el-table-column prop="" label="Bán kính">
+        <template #default="scope">
+          {{ scope.row.radius ?? scope.row.Area.radius }}
+        </template>
+      </el-table-column>
+      <el-table-column label="Hành động">
+        <template #default="scope">
+          <el-button type="danger" plain round @click="onDeleteArea(scope.row.ID)">
+            Xoá
+          </el-button>
+        </template>
+      </el-table-column>
+    </el-table>
+    <div class="flex justify-end">
+      <el-pagination
+        v-model:current-page="page" v-model:page-size="pageSize" :page-sizes="[20, 50, 100, 500]"
+        :size="size" :background="true" layout="total, sizes, prev, pager, next, jumper" :total="total"
+        @size-change="handleSizeChange" @current-change="handleCurrentChange"
+      />
+    </div>
+    <el-drawer
+      v-model="dialogFormVisible" destroy-on-close size="800" :show-close="false"
+      :before-close="closeDialog"
+    >
+      <template #header>
+        <div class="flex justify-between items-center">
+          <span class="text-lg">{{ type === 'create' ? 'Thêm mới' : 'Chỉnh sửa' }}</span>
+          <div>
+            <el-button type="primary" @click="enterDialog">
+              Đồng ý
+            </el-button>
+            <el-button @click="dialogFormVisible = false">
+              Hủy
+            </el-button>
+          </div>
+        </div>
+      </template>
+
+      <el-form ref="elFormRef" :model="formData" label-position="top" :rules="rules" label-width="80px">
+        <el-form-item label="Tên khu vực:" prop="areaId">
+          <el-select
+            v-model="formData.areaId" placeholder="Chọn khu vực" clearable filterable
+            @change="onSelectChange"
+          >
+            <el-option v-for="item in areaOptions" :key="item.ID" :label="item.name" :value="item.ID" />
+          </el-select>
+        </el-form-item>
+        <el-form-item label="Bán kính:" prop="radius">
+          <el-input-number v-model="formData.radius" style="width:100%" :precision="2" :clearable="true" />
+        </el-form-item>
+      </el-form>
+    </el-drawer>
+  </div>
 </template>
 
 <script setup>
@@ -63,7 +77,7 @@ import {
     findAttendanceArea
 } from '@/api/checkins/attendance'
 import { ref } from 'vue'
-import { ElMessage } from 'element-plus';
+import { ElMessage, ElMessageBox } from 'element-plus';
 
 const props = defineProps(
     {
@@ -108,20 +122,34 @@ const getAreaListData = async () => {
         tableData.value = table.data
         total.value = table.data.total
     }
-  
+
 }
 
 getAreaListData();
 
 const onDeleteArea = async (id) => {
-    const res = await deleteAttendanceArea({ id: id })
-    if (res.code === 0) {
+    if (!id) return
+
+    ElMessageBox.confirm('Khi xoá khu vực đồng thời cũng sẽ xoá điều kiện tham chiếu', 'Cảnh báo', {
+        confirmButtonText: 'Đồng ý',
+        cancelButtonText: 'Hủy',
+        type: 'warning'
+    }).then(async() => {
+        const res = await deleteAttendanceArea({ id: id })
+        if (res.code === 0) {
+            ElMessage({
+                type: 'success',
+                message: 'Xoá thành công'
+            })
+            getAreaListData();
+        }
+    }).catch(() => {
         ElMessage({
-            type: 'success',
-            message: 'Xoá thành công'
+            type: 'info',
+            message: 'Đã hủy'
         })
-        getAreaListData();
-    }
+    })
+
 }
 
 const addNewArea = () => {
